+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## [2023.2.12]
+
+### `ImpEx` enhancements
+- Added colorization for `odd` and `even` user rights value lines [#802](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/802)
+
 ## [2023.2.11]
 
 ### `Project Import` enhancements